refactor: Order 스키마 통합
